requirements.txt
torch~=2.0.1
torchsummary~=1.5.1
torchvision~=0.15.2
numpy
scikit-learn~=1.3.0
tqdm~=4.65.0
matplotlib~=3.7.1
Pillow~=10.0.0
scikit-image~=0.21.0
opencv-python~=4.8.0.74